+++ packages/cadl-ranch-coverage-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @azure-tools/cadl-ranch-api
 
+## 0.2.2
+
+### Patch Changes
+
+- 3a6b67d: add tsv, ssv, pipes query collection format test
+
 ## 0.2.1
 
 ### Patch Changes